buffer handling modernized

+++ ipd.h
@@ -26,7 +26,7 @@
 #define IPD_DIR "/tmp/ipd"
 #define IPD_MAX_APP_NAME_LEN 31
 #define IPD_PORT 12193
-#define IPD_BUFSIZ 2048
+#define IPD_MAXPUB 8192
 
 
 struct ipd
@@ -35,7 +35,13 @@ struct ipd
   int fd;
   struct ev_loop *loop;
   void *ud;
-  int (*cb)(void *,char *,int,unsigned *);
+  union
+  {
+    int (*rcb)(void *,char *,int,char **);
+    int (*bcb)(void *,char *,int);
+  } u;
+  char *rxbuf;
+  int rxblen;
   char app[IPD_MAX_APP_NAME_LEN+1];
 };
 
@@ -47,14 +53,14 @@ struct ipd
  *    app  - application name, must be unique system wide
  *    loop - libev loop
  *    cb   - callback for requests
- *           int cb(void *ud, char *req, int len, unsigned *replylen)
+ *           int cb(void *ud, char *req, int len, char **reply)
  *             ud       - userdata
- *             req      - request as c string, reply on return
+ *             req      - request as c string
  *             len      - length of request including terminator
- *             replylen - length of reply (<=len)
+ *             reply    - reply
  *    ud   - userdata, passed to cb
  */
-static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop, int (*cb)(void *,char *,int, unsigned *), void *ud);
+static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop, int (*cb)(void *,char *,int,char **), void *ud);
 
 /*
  *  unregister application
@@ -82,7 +88,7 @@ static inline int ipd_pub(const char *msg);
  *             len      - length of request including terminator
  *    ud   - userdata passed to cb
  */
-static inline int ipd_sub(struct ipd *ipd, struct ev_loop *loop, int (*cb)(void *,char *,int, unsigned *), void *ud);
+static inline int ipd_sub(struct ipd *ipd, struct ev_loop *loop, int (*cb)(void *,char *,int), void *ud);
 
 /*
  *  send an rpc request for an application
@@ -99,17 +105,18 @@ static inline void ipd_process_command_cb(struct ev_loop* loop, struct ev_io *io
 {
   struct ipd *ipd=(struct ipd *)io;
   struct sockaddr_un ca={0};
-  char buf[IPD_BUFSIZ];
+  char buf[1];
   int n;
   socklen_t cl=sizeof(struct sockaddr_un);
-  unsigned rs=0;
+  char *rpl=0;
   
   ca.sun_family=AF_UNIX;
-  if(-1!=(n=recvfrom(ipd->fd,buf,sizeof(buf),0,(struct sockaddr *)&ca,&cl))&&0!=ipd->cb)
+  if(-1!=(n=recvfrom(ipd->fd,buf,sizeof(buf),MSG_TRUNC|MSG_PEEK,(struct sockaddr *)&ca,&cl))&&0!=ipd->u.rcb)
   {
-    ipd->cb(ipd->ud,buf,n,&rs);
-    if(rs==0) buf[rs++]=0;
-    sendto(ipd->fd,buf,rs,0,(struct sockaddr *)&ca,cl);
+    if(ipd->rxblen<n) ipd->rxbuf=realloc(ipd->rxbuf,ipd->rxblen=n);
+    recvfrom(ipd->fd,ipd->rxbuf,n,0,0,0);
+    ipd->u.rcb(ipd->ud,ipd->rxbuf,n,&rpl);
+    if(0!=rpl) sendto(ipd->fd,rpl,strlen(rpl)+1,0,(struct sockaddr *)&ca,cl);
   }
 }
 
@@ -122,12 +129,17 @@ static inline void ipd_unreg(struct ipd *ipd)
   {
     ev_io_stop(ipd->loop,&ipd->io);
     close(ipd->fd);
-    snprintf(nam,sizeof(nam),"%s/%s",IPD_DIR,ipd->app);
-    unlink(nam);
+    if(0!=ipd->rxbuf) free(ipd->rxbuf);
+    if(0!=ipd->app[0])
+    {
+      snprintf(nam,sizeof(nam),"%s/%s",IPD_DIR,ipd->app);
+      unlink(nam);
+    }
+    bzero(ipd,sizeof(struct ipd));
   }
 }
 
-static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop, int (*cb)(void *,char *,int, unsigned *), void *ud)
+static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop, int (*cb)(void *,char *,int,char **), void *ud)
 {
   int ret=-1;
   char nam[sizeof(IPD_DIR)+1+IPD_MAX_APP_NAME_LEN+1];
@@ -138,8 +150,10 @@ static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop
   {
     ipd->fd=-1;
     ipd->loop=loop;
-    ipd->cb=cb;
+    ipd->u.rcb=cb;
     ipd->ud=ud;
+    ipd->rxbuf=0;
+    ipd->rxblen=0;
     strncpy(ipd->app,app,sizeof(ipd->app));
     snprintf(nam,sizeof(nam),"%s/%s",IPD_DIR,app);
     if(-1!=(ipd->fd=socket(AF_UNIX,SOCK_DGRAM,0)))
@@ -161,19 +175,19 @@ static inline int ipd_reg(struct ipd *ipd, const char *app, struct ev_loop *loop
 
 static inline int ipd_pub(const char *msg)
 {
-  int ret=-1;
+  int ret=-1,l;
   int fd;
   const int bc=1;
   struct sockaddr_in ad={0};
   
-  if(NULL!=msg)
+  if(NULL!=msg&&(l=strlen(msg))<IPD_MAXPUB)
   {
     fd=socket(AF_INET,SOCK_DGRAM,0);
     setsockopt(fd,SOL_SOCKET,SO_BROADCAST,&bc,sizeof(bc));
     ad.sin_family=AF_INET;
     ad.sin_port=(in_port_t)htons(IPD_PORT);
     ad.sin_addr.s_addr=inet_addr("127.255.255.255");
-    ret=sendto(fd,msg,strlen(msg)+1,0,(struct sockaddr *)&ad,sizeof(ad));
+    ret=sendto(fd,msg,l+1,0,(struct sockaddr *)&ad,sizeof(ad));
   }
   
   return(ret);
@@ -182,16 +196,15 @@ static inline int ipd_pub(const char *msg)
 static inline void ipd_sub_cb(struct ev_loop* loop, struct ev_io *io, int r)
 {
   struct ipd *ipd=(struct ipd *)io;
-  char buf[IPD_BUFSIZ];
+  char buf[IPD_MAXPUB];
   int n;
   
-  if(-1!=(n=recvfrom(ipd->fd,buf,sizeof(buf),0,0,0))&&0!=ipd->cb)
-  {
-    ipd->cb(ipd->ud,buf,n,0);
-  }
+  if(-1!=(n=recvfrom(ipd->fd,buf,sizeof(buf),0,0,0))&&0!=ipd->u.bcb) ipd->u.bcb(ipd->ud,buf,n);
 }
 
-static inline int ipd_sub(struct ipd *ipd, struct ev_loop *loop, int (*cb)(void *,char *,int, unsigned *), void *ud)
+#define ipd_unsub(ipd) ipd_unreg(ipd)
+
+static inline int ipd_sub(struct ipd *ipd, struct ev_loop *loop, int (*cb)(void *,char *,int), void *ud)
 {
   int ret=-1;
   const int enable=1;
@@ -199,13 +212,18 @@ static inline int ipd_sub(struct ipd *ipd, struct ev_loop *loop, int (*cb)(void 
 
   if(0!=ipd&&0!=loop&&0!=cb)
   {
+    ipd->loop=loop;
+    ipd->rxbuf=0;
+    ipd->rxblen=0;
+    ipd->app[0]=0;
     ipd->fd=socket(PF_INET,SOCK_DGRAM,0);
     setsockopt(ipd->fd,SOL_SOCKET,SO_REUSEADDR,&enable,sizeof(int));
+    setsockopt(ipd->fd,SOL_SOCKET,SO_REUSEPORT,&enable,sizeof(int));
     ad.sin_family=AF_INET;
     ad.sin_port=htons(IPD_PORT);
     ad.sin_addr.s_addr=INADDR_ANY;
     bind(ipd->fd,(struct sockaddr*)&ad,sizeof(ad));
-    ipd->cb=cb;
+    ipd->u.bcb=cb;
     ipd->ud=ud;
     ev_io_init(&ipd->io,ipd_sub_cb,ipd->fd,EV_READ);
     ev_io_start(loop,&ipd->io);
